+++ dds-agent/src/CommanderChannel.cpp
@@ -331,3 +331,8 @@ bool CCommanderChannel::on_cmdACTIVATE_AGENT(SCommandAttachmentImpl<cmdACTIVATE_
 
     return true;
 }
+
+bool CCommanderChannel::on_cmdUPDATE_KEY(SCommandAttachmentImpl<cmdUPDATE_KEY>::ptr_t _attachment)
+{
+    return true;
+}